Add files via upload
In this file I had write code about output possibilities in javascript. first one is document.write(); in which we will get an output in output screen. second one is console.log(); in which we will get output message in web console. third is alert(); it gives an output in popup alert. and the last one is document.getElementById(); in which we will get output for an particular tag or particular id.